+++ src/app/main/page.tsx
@@ -1,8 +1,8 @@
 "use client"
 import React, { useEffect, useState } from "react"
 import { Container } from "@mui/material"
-import useGetOgsm from "@/hooks/useGetOgsm"
-import useAuth from "@/hooks/useAuth"
+import useGetOgsm from "@/hooks/ogsm/useGetOgsm"
+import useAuth from "@/hooks/common/useAuth"
 import { getAuth } from "firebase/auth"
 import "react-toastify/dist/ReactToastify.css"
 import Header from "@/components/features/main/Header"
