+++ src/components/features/main/Header.tsx
@@ -1,6 +1,6 @@
-import useAuth from "@/hooks/useAuth"
-import useLogin from "@/hooks/useLogin"
-import useLogout from "@/hooks/useLogout"
+import useAuth from "@/hooks/common/useAuth"
+import useLogin from "@/hooks/common/useLogin"
+import useLogout from "@/hooks/common/useLogout"
 import { Button } from "@mui/material"
 import { GoogleAuthProvider } from "firebase/auth"
 import React from "react"
